% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_reqs.r
\name{apiReq}
\alias{apiReq}
\title{Formats an movebank api request}
\usage{
apiReq(params)
}
\arguments{
\item{params}{\code{list} List in which item name is the api parameter name, and value is the value to send to api}
}
\value{
\code{character} A request url
}
\description{
Formats an movebank api request
}
