% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_movebank_study.r
\name{downloadStudy}
\alias{downloadStudy}
\title{Downloads study metadata using movebank api}
\usage{
downloadStudy(studyid, userid)
}
\arguments{
\item{studyid}{\code(integer) The id of the study}

\item{userid}{\code(string) Movebank user id}
}
\value{
\code(tibble) A dataframe of information about the study
}
\description{
Downloads study metadata using movebank api
}
\examples{
downloadStudy(<id>,"ben.s.carlson")
}
