% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_movebank_event.r
\name{eventAPICall}
\alias{eventAPICall}
\title{Creates an movebank api call
Usage example: eventAPICall(study_id=9648615)}
\usage{
eventAPICall(study_id, start = NULL, end = NULL)
}
\description{
TODO: filter by visible!=FALSE.
visible: Determines whether an event is visible on the Movebank Search map. Values are calculated automatically, with FALSE indicating that the event has been marked as an outlier by manually marked outlier or algorithm marked outlier. Allowed values are TRUE or FALSE.
eobs_temperature',
}
