% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.r
\name{formatTs}
\alias{formatTs}
\title{Formats a POSIXct object into a format required by movebank api}
\usage{
formatTs(ts)
}
\arguments{
\item{ts}{\code{POSIXct} Timestamp to be converted to movebank api format}
}
\value{
\code{character} Formatted timestamp
}
\description{
Formats a POSIXct object into a format required by movebank api
}
