% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getDeployment}
\alias{getDeployment}
\title{Downloads deployment data using movebank api}
\usage{
getDeployment(studyid, params = list(), ...)
}
\arguments{
\item{studyid}{\code{integer} The id of the study. Must only be a single id (unlike getStudy).}
}
\value{
Deployment data
}
\description{
Downloads deployment data using movebank api
}
\examples{
getDeployment(123413)
}
