% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getEvent}
\alias{getEvent}
\title{Downloads event data for a study using movebank api}
\usage{
getEvent(studyid, userid)
}
\arguments{
\item{studyid}{The id of the study}

\item{userid}{Movebank user id}
}
\value{
Events for the study
}
\description{
Downloads event data for a study using movebank api
}
\examples{
getEvent(123413,"ben.s.carlson")
}
