% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getIndividual}
\alias{getIndividual}
\title{Downloads individual data using movebank api}
\usage{
getIndividual(studyid, params = list(), urlonly = FALSE)
}
\arguments{
\item{studyid}{\code{integer} The id of the study. Must only be a single id (unlike getStudy).}
}
\value{
Information about the study
}
\description{
Downloads individual data using movebank api
}
\examples{
getIndividual(123413)
}
