% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getSensor}
\alias{getSensor}
\title{Downloads sensor data using movebank api}
\usage{
getSensor(studyid, params = list(), ...)
}
\arguments{
\item{studyid}{\code{integer} The id of the study. Must only be a single id.}
}
\value{
Sensor data
}
\description{
Downloads sensor data using movebank api
}
\examples{
getSensor(123413)
}
