% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getStudy}
\alias{getStudy}
\title{Downloads study data using movebank api}
\usage{
getStudy(studyid, params = list(), urlonly = FALSE)
}
\arguments{
\item{studyid}{\code{integer} The id of the study. Can be a vector of ids.}
}
\value{
Information about the study
}
\description{
Downloads study data using movebank api
}
\examples{
getStudy(123413)
}
