% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.r
\name{setAuth}
\alias{setAuth}
\title{Sets the userid for the r session}
\usage{
setAuth(userid, pass)
}
\arguments{
\item{userid}{\code{character} Movebank user id}

\item{pass}{\code{character} Movebank password}
}
\description{
Sets the userid for the r session
}
