% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getEvent}
\alias{getEvent}
\title{Downloads event data for a study using movebank api}
\usage{
getEvent(
  studyid,
  attributes = NULL,
  sensor_type_id = NULL,
  ts_start = NULL,
  ts_end = NULL,
  params = list(),
  ...
)
}
\arguments{
\item{studyid}{The id of the study}

\item{attributes}{\code{character} Vector of desired attributes. To get all attributes use attributes=all}

\item{ts_start}{\code{character} Assumes UTC}

\item{ts_end}{\code{character} Assumes UTC}
}
\value{
Events for the study
}
\description{
Downloads event data for a study using movebank api
}
\examples{
getEvent(123413)
}
