% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getMvData}
\alias{getMvData}
\title{TODO: do some checking of outputs to make sure I got a good response}
\usage{
getMvData(apiReq, userid)
}
\description{
TODO: do some checking of outputs to make sure I got a good response
}
