% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getMvData}
\alias{getMvData}
\title{Gets movebank data based on api request}
\usage{
getMvData(
  apiReq,
  accept_license = FALSE,
  handle = NULL,
  save_as = NULL,
  clean = TRUE
)
}
\arguments{
\item{apiReq}{\code{character} URL for API request.}

\item{accept_license}{\code{boolean} Set to TRUE to use md5 method to accept license terms over api.}

\item{handle}{\code{handle} Handle object used in httr. Mainly for testing purposes to start with blank session. Use as handle=handle('')}

\item{save_as}{\code{character} Save response directly to disk.}

\item{clean}{\code{character} Defaults to true. Removes line breaks and trims whitespace from all fields}
}
\value{
\code{data.frame} Data from API request. If saving results to disk, returns TRUE.
}
\description{
Gets movebank data based on api request
}
\examples{
apiReq <- https://www.movebank.org/movebank/service/direct-read?entity_type=study&study_id=2911040
getMvData(apiReq,accept_license=TRUE)

}
