% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.r
\name{getStudy}
\alias{getStudy}
\title{Downloads study metadata using movebank api}
\usage{
getStudy(studyid, userid)
}
\arguments{
\item{studyid}{The id of the study}

\item{userid}{Movebank user id}
}
\value{
Information about the study
}
\description{
Downloads study metadata using movebank api
}
\examples{
getStudy(123413,"ben.s.carlson")
}
